Create README.md
This is the customized class teacher dashboard. It is made by me using Microsoft Excel. It shows the details of my whole class. Like mode of education(Day Scholar or Hosteler), their blood groups,no of boys and girls present in the class, availability of the laptops for students and GATE aspirants etc.